- a little fix on the roles checkbox for allow barangay

+++ api/delete_role.php
@@ -9,16 +9,14 @@ require_once '../db/db.php';
 try {
   if (!$_SERVER['REQUEST_METHOD'] == 'POST') throw new Exception('Invalid request.');
   if (empty($_POST['id'])) throw new Exception('id not found.');
-  if (empty($_POST['barPerms'])) throw new Exception('barPerms not found.');
-  if (empty($_POST['genPerms'])) throw new Exception('genPerms not found.');
 
   // get pdo
   global $pdo;
 
   // assign id
   $id  = trim($_POST['id']);
-  $barPermID  = trim($_POST['barPerms']);
-  $genPermID  = trim($_POST['genPerms']);
+  $barPermID  = empty($_POST['barPerms']) ? null : $_POST['barPerms'];
+  $genPermID  = empty($_POST['genPerms']) ? null : $_POST['genPerms'];
 
   // delete role  (delete first because of foreign key)
   $sql = 'delete from roles where id = :id';
@@ -26,14 +24,18 @@ try {
   $query->execute([':id' => $id]);
 
   // delete gen permissions
-  $sql = 'delete from permissions where id = :gen_perms';
-  $query = $pdo->prepare($sql);
-  $query->execute([':gen_perms' => $genPermID]);
+  if ($genPermID) {
+    $sql = 'delete from permissions where id = :gen_perms';
+    $query = $pdo->prepare($sql);
+    $query->execute([':gen_perms' => $genPermID]);
+  }
 
   // delete bar permissions
-  $sql = 'delete from permissions where id = :bar_perms';
-  $query = $pdo->prepare($sql);
-  $query->execute([':bar_perms' => $barPermID]);
+  if ($barPermID) {
+    $sql = 'delete from permissions where id = :bar_perms';
+    $query = $pdo->prepare($sql);
+    $query->execute([':bar_perms' => $barPermID]);
+  }
 
   echo json_encode('Success');
 } catch (\Throwable $th) {
